+++ src/Catalyst.jl
@@ -94,7 +94,7 @@ export dependants, dependents, substoichmat, prodstoichmat, netstoichmat
 export isautonomous
 export reactionrates
 export set_default_noise_scaling
-export make_rre_ode, make_cle_sde, make_sck_jump, make_rre_algeqs
+export make_rre_ode, make_cle_sde, make_sck_jump, make_rre_algeqs, make_hybrid_model
 
 # depreciated functions to remove in future releases
 export params, numparams